+++ Code/BuyOrderCalulator.Web.Server/Models/ToClient/OrderViewModel.cs
@@ -19,7 +19,7 @@ namespace BuyOrderCalc.Web.Server.Models.ToClient
     public class OrderItemModel
     {
         public int Quantity { get; set; }
-        public int FixedUnitPrice { get; set; }
+        public double FixedUnitPrice { get; set; }
         public double FixedCorpCreditPercent { get; set; }
 
         public int ItemId { get; set; }
